+++ simulations/vip-507/arbitrumone.ts
@@ -5,12 +5,12 @@ import { expectEvents } from "src/utils";
 import { forking, testForkedNetworkVipCommands } from "src/vip-framework";
 import { checkTwoKinksInterestRateIL } from "src/vip-framework/checks/interestRateModel";
 
-import vip506, {
+import vip507, {
   ARB_vUSDC_Core,
   ARB_vUSDC_Core_IRM,
   ARB_vUSDT_Core,
   ARB_vUSDT_Core_IRM,
-} from "../../vips/vip-506/bscmainnet";
+} from "../../vips/vip-507/bscmainnet";
 import VTOKEN_ABI from "./abi/VToken.json";
 
 export const SECONDS_PER_YEAR = 31_536_000; // seconds per year
@@ -45,7 +45,7 @@ forking(342077800, async () => {
     );
   });
 
-  testForkedNetworkVipCommands("VIP 506", await vip506(), {
+  testForkedNetworkVipCommands("VIP 507", await vip507(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(txResponse, [VTOKEN_ABI], ["NewMarketInterestRateModel"], [2]);
     },
