+++ simulations/vip-507/basemainnet.ts
@@ -1,28 +1,33 @@
 import { expect } from "chai";
-import { BigNumber } from "ethers";
+import { BigNumber, Contract } from "ethers";
 import { ethers } from "hardhat";
 import { expectEvents } from "src/utils";
 import { forking, testForkedNetworkVipCommands } from "src/vip-framework";
 import { checkInterestRate, checkTwoKinksInterestRateIL } from "src/vip-framework/checks/interestRateModel";
 
-import vip506, { UNI_vUSDC_Core, UNI_vUSDC_Core_IRM } from "../../vips/vip-506/bsctestnet";
+import vip507, { BASE_vUSDC_Core, BASE_vUSDC_Core_IRM } from "../../vips/vip-507/bscmainnet";
 import VTOKEN_ABI from "./abi/VToken.json";
 
-export const SECONDS_PER_YEAR = 31_536_000;
+export const SECONDS_PER_YEAR = 31_536_000; // seconds per year
 
-forking(21502211, async () => {
+forking(30907412, async () => {
+  let vUSDC_Core: Contract;
+
+  before(async () => {
+    vUSDC_Core = new ethers.Contract(BASE_vUSDC_Core, VTOKEN_ABI, ethers.provider);
+  });
+
   describe("Pre-VIP behaviour", async () => {
     it("check IRM address", async () => {
-      const vUSDC_Core = new ethers.Contract(UNI_vUSDC_Core, VTOKEN_ABI, ethers.provider);
-      expect(await vUSDC_Core.interestRateModel()).to.equals("0x2f8263A8833C21b6B72AC55951756B41D4607e3F");
+      expect(await vUSDC_Core.interestRateModel()).to.equals("0xB7FCED42486F8DB8155fD509e726F9604fCdd41F");
     });
 
     checkInterestRate(
-      "0x2f8263A8833C21b6B72AC55951756B41D4607e3F",
+      "0xB7FCED42486F8DB8155fD509e726F9604fCdd41F",
       "USDC_CORE",
       {
         base: "0",
-        multiplier: "0.0875",
+        multiplier: "0.08",
         jump: "2.5",
         kink: "0.8",
       },
@@ -30,7 +35,7 @@ forking(21502211, async () => {
     );
   });
 
-  testForkedNetworkVipCommands("VIP 506", await vip506(), {
+  testForkedNetworkVipCommands("VIP 507", await vip507(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(txResponse, [VTOKEN_ABI], ["NewMarketInterestRateModel"], [1]);
     },
@@ -38,12 +43,11 @@ forking(21502211, async () => {
 
   describe("Post-VIP behavior", () => {
     it("check IRM address", async () => {
-      const vUSDC_Core = new ethers.Contract(UNI_vUSDC_Core, VTOKEN_ABI, ethers.provider);
-      expect(await vUSDC_Core.interestRateModel()).to.equals(UNI_vUSDC_Core_IRM);
+      expect(await vUSDC_Core.interestRateModel()).to.equals(BASE_vUSDC_Core_IRM);
     });
 
     checkTwoKinksInterestRateIL(
-      UNI_vUSDC_Core_IRM,
+      BASE_vUSDC_Core_IRM,
       "USDC_Core",
       {
         base: "0",
