+++ simulations/vip-507/basesepolia.ts
@@ -1,32 +1,29 @@
 import { expect } from "chai";
-import { BigNumber, Contract } from "ethers";
+import { BigNumber } from "ethers";
 import { ethers } from "hardhat";
 import { expectEvents } from "src/utils";
 import { forking, testForkedNetworkVipCommands } from "src/vip-framework";
 import { checkInterestRate, checkTwoKinksInterestRateIL } from "src/vip-framework/checks/interestRateModel";
 
-import vip506, { UNI_vUSDC_Core, UNI_vUSDC_Core_IRM } from "../../vips/vip-506/bscmainnet";
+import vip506, { BASE_vUSDC_Core, BASE_vUSDC_Core_IRM } from "../../vips/vip-507/bsctestnet";
 import VTOKEN_ABI from "./abi/VToken.json";
 
 export const SECONDS_PER_YEAR = 31_536_000; // seconds per year
-let vUSDC_Core: Contract;
 
-forking(17858329, async () => {
-  before(async () => {
-    vUSDC_Core = new ethers.Contract(UNI_vUSDC_Core, VTOKEN_ABI, ethers.provider);
-  });
+forking(26290932, async () => {
   describe("Pre-VIP behaviour", async () => {
     it("check IRM address", async () => {
-      expect(await vUSDC_Core.interestRateModel()).to.equals("0x2DAb0B51d7d899b66d4F16EA1c0cEB1767863523");
+      const vUSDC_Core = new ethers.Contract(BASE_vUSDC_Core, VTOKEN_ABI, ethers.provider);
+      expect(await vUSDC_Core.interestRateModel()).to.equals("0x37a3e7951795d3aF33D606A9591B8EBd5F84DA08");
     });
 
     checkInterestRate(
-      "0x2DAb0B51d7d899b66d4F16EA1c0cEB1767863523",
+      "0x37a3e7951795d3aF33D606A9591B8EBd5F84DA08",
       "USDC_CORE",
       {
         base: "0",
-        multiplier: "0.125",
-        jump: "3",
+        multiplier: "0.075",
+        jump: "2.5",
         kink: "0.8",
       },
       BigNumber.from(SECONDS_PER_YEAR),
@@ -41,11 +38,12 @@ forking(17858329, async () => {
 
   describe("Post-VIP behavior", () => {
     it("check IRM address", async () => {
-      expect(await vUSDC_Core.interestRateModel()).to.equals(UNI_vUSDC_Core_IRM);
+      const vUSDC_Core = new ethers.Contract(BASE_vUSDC_Core, VTOKEN_ABI, ethers.provider);
+      expect(await vUSDC_Core.interestRateModel()).to.equals(BASE_vUSDC_Core_IRM);
     });
 
     checkTwoKinksInterestRateIL(
-      UNI_vUSDC_Core_IRM,
+      BASE_vUSDC_Core_IRM,
       "USDC_Core",
       {
         base: "0",
@@ -54,7 +52,7 @@ forking(17858329, async () => {
         multiplier2: "0.7",
         base2: "0",
         kink2: "0.9",
-        jump: "3",
+        jump: "2.5",
       },
       BigNumber.from(SECONDS_PER_YEAR),
     );
