+++ simulations/vip-507/ethereum.ts
@@ -5,56 +5,55 @@ import { expectEvents } from "src/utils";
 import { forking, testForkedNetworkVipCommands } from "src/vip-framework";
 import { checkInterestRate, checkTwoKinksInterestRateIL } from "src/vip-framework/checks/interestRateModel";
 
-import vip506, {
-  SEPOLIA_vDAI_Core,
-  SEPOLIA_vDAI_Core_IRM,
-  SEPOLIA_vFRAX_Core,
-  SEPOLIA_vFRAX_Core_IRM,
-  SEPOLIA_vTUSD_Core,
-  SEPOLIA_vTUSD_Core_IRM,
-  SEPOLIA_vUSDC_Core,
-  SEPOLIA_vUSDC_Core_IRM,
-  SEPOLIA_vUSDC_Ethena,
-  SEPOLIA_vUSDC_Ethena_IRM,
-  SEPOLIA_vUSDS_Core,
-  SEPOLIA_vUSDS_Core_IRM,
-  SEPOLIA_vUSDT_Core,
-  SEPOLIA_vUSDT_Core_IRM,
-  SEPOLIA_vcrvUSD_Core,
-  SEPOLIA_vcrvUSD_Core_IRM,
-  SEPOLIA_vcrvUSD_Curve,
-  SEPOLIA_vcrvUSD_Curve_IRM,
-} from "../../vips/vip-506/bsctestnet";
+import vip507, {
+  ETH_vDAI_Core,
+  ETH_vDAI_Core_IRM,
+  ETH_vFRAX_Core,
+  ETH_vFRAX_Core_IRM,
+  ETH_vTUSD_Core,
+  ETH_vTUSD_Core_IRM,
+  ETH_vUSDC_Core,
+  ETH_vUSDC_Core_IRM,
+  ETH_vUSDC_Ethena,
+  ETH_vUSDC_Ethena_IRM,
+  ETH_vUSDS_Core,
+  ETH_vUSDS_Core_IRM,
+  ETH_vUSDT_Core,
+  ETH_vUSDT_Core_IRM,
+  ETH_vcrvUSD_Core,
+  ETH_vcrvUSD_Core_IRM,
+  ETH_vcrvUSD_Curve,
+  ETH_vcrvUSD_Curve_IRM,
+} from "../../vips/vip-507/bscmainnet";
 import VTOKEN_ABI from "./abi/VToken.json";
 
 export const ETH_BLOCKS_PER_YEAR = 2_628_000; // assuming a block is mined every 12 seconds
-export const OLD_ETH_BLOCK_PER_YER = 2_252_571;
 
 const oldIRMs = [
-  [SEPOLIA_vUSDS_Core, "0xBd20E6922A5b1e20B5e611700f1Cab7c177C35De"],
-  [SEPOLIA_vUSDC_Core, "0xA03205bC635A772E533E7BE36b5701E331a70ea3"],
-  [SEPOLIA_vcrvUSD_Core, "0xA03205bC635A772E533E7BE36b5701E331a70ea3"],
-  [SEPOLIA_vFRAX_Core, "0xfBa27be0766acb9a60d2cede3d4293193f3b749d"],
-  [SEPOLIA_vDAI_Core, "0xB778B803a72Af7ac9Ac8277bA87B405cE91741B9"],
-  [SEPOLIA_vUSDT_Core, "0xA03205bC635A772E533E7BE36b5701E331a70ea3"],
-  [SEPOLIA_vTUSD_Core, "0xfBa27be0766acb9a60d2cede3d4293193f3b749d"],
-  [SEPOLIA_vcrvUSD_Curve, "0xA03205bC635A772E533E7BE36b5701E331a70ea3"],
-  [SEPOLIA_vUSDC_Ethena, "0x7a381Fa8B502ac3870D05ECfb321CFEf1818A218"],
+  [ETH_vUSDS_Core, "0x322072b84434609ff64333A125516055B5B4405F"],
+  [ETH_vUSDC_Core, "0x4A786e4653Ff7DBA74D6dA0861350F233f2dA73b"],
+  [ETH_vcrvUSD_Core, "0x837996e7d74222965ACf1fdd478926e07336a291"],
+  [ETH_vFRAX_Core, "0x675b3dF06a5F3A7d2f04e7852Dbf8f8d40959Ca9"],
+  [ETH_vDAI_Core, "0x675b3dF06a5F3A7d2f04e7852Dbf8f8d40959Ca9"],
+  [ETH_vUSDT_Core, "0x4A786e4653Ff7DBA74D6dA0861350F233f2dA73b"],
+  [ETH_vTUSD_Core, "0x675b3dF06a5F3A7d2f04e7852Dbf8f8d40959Ca9"],
+  [ETH_vcrvUSD_Curve, "0x837996e7d74222965ACf1fdd478926e07336a291"],
+  [ETH_vUSDC_Ethena, "0x786ed1931Fa76a70B1259c42B0aC106B1E08485c"],
 ];
 
 const newIRMs = [
-  [SEPOLIA_vUSDS_Core, SEPOLIA_vUSDS_Core_IRM],
-  [SEPOLIA_vUSDC_Core, SEPOLIA_vUSDC_Core_IRM],
-  [SEPOLIA_vcrvUSD_Core, SEPOLIA_vcrvUSD_Core_IRM],
-  [SEPOLIA_vFRAX_Core, SEPOLIA_vFRAX_Core_IRM],
-  [SEPOLIA_vDAI_Core, SEPOLIA_vDAI_Core_IRM],
-  [SEPOLIA_vUSDT_Core, SEPOLIA_vUSDT_Core_IRM],
-  [SEPOLIA_vTUSD_Core, SEPOLIA_vTUSD_Core_IRM],
-  [SEPOLIA_vcrvUSD_Curve, SEPOLIA_vcrvUSD_Curve_IRM],
-  [SEPOLIA_vUSDC_Ethena, SEPOLIA_vUSDC_Ethena_IRM],
+  [ETH_vUSDS_Core, ETH_vUSDS_Core_IRM],
+  [ETH_vUSDC_Core, ETH_vUSDC_Core_IRM],
+  [ETH_vcrvUSD_Core, ETH_vcrvUSD_Core_IRM],
+  [ETH_vFRAX_Core, ETH_vFRAX_Core_IRM],
+  [ETH_vDAI_Core, ETH_vDAI_Core_IRM],
+  [ETH_vUSDT_Core, ETH_vUSDT_Core_IRM],
+  [ETH_vTUSD_Core, ETH_vTUSD_Core_IRM],
+  [ETH_vcrvUSD_Curve, ETH_vcrvUSD_Curve_IRM],
+  [ETH_vUSDC_Ethena, ETH_vUSDC_Ethena_IRM],
 ];
 
-forking(8412432, async () => {
+forking(22595258, async () => {
   describe("Pre-VIP behaviour", async () => {
     it("check IRM address", async () => {
       for (const [market, expectedIRM] of oldIRMs) {
@@ -64,8 +63,8 @@ forking(8412432, async () => {
     });
 
     checkInterestRate(
-      "0xBd20E6922A5b1e20B5e611700f1Cab7c177C35De",
-      "USDS_CORE",
+      "0x322072b84434609ff64333A125516055B5B4405F",
+      "USDS_Core",
       {
         base: "0",
         multiplier: "0.15625",
@@ -75,24 +74,27 @@ forking(8412432, async () => {
       BigNumber.from(ETH_BLOCKS_PER_YEAR),
     );
 
-    checkInterestRate(
-      "0xA03205bC635A772E533E7BE36b5701E331a70ea3",
-      "USDC_CORE_USDT_CORE_vcrvUSD_CORE_vcrv_USD_Curve",
-      {
-        base: "0",
-        multiplier: "0.07",
-        jump: "0.8",
-        kink: "0.8",
-      },
-      BigNumber.from(OLD_ETH_BLOCK_PER_YER),
-    );
-
-    checkInterestRate(
-      "0xfBa27be0766acb9a60d2cede3d4293193f3b749d",
-      "FRAX_Core_TUSD_Core",
+    checkTwoKinksInterestRateIL(
+      "0x4A786e4653Ff7DBA74D6dA0861350F233f2dA73b",
+      "USDC_Core_USDT_Core",
       {
         base: "0",
         multiplier: "0.15",
+        kink1: "0.8",
+        multiplier2: "0.9",
+        base2: "0",
+        kink2: "0.9",
+        jump: "3.0",
+      },
+      BigNumber.from(ETH_BLOCKS_PER_YEAR),
+    );
+
+    checkInterestRate(
+      "0x837996e7d74222965ACf1fdd478926e07336a291",
+      "crvUSD_CORE_crvUSD_Curve",
+      {
+        base: "0",
+        multiplier: "0.2",
         jump: "2.5",
         kink: "0.8",
       },
@@ -100,11 +102,11 @@ forking(8412432, async () => {
     );
 
     checkInterestRate(
-      "0xB778B803a72Af7ac9Ac8277bA87B405cE91741B9",
-      "DAI_CORE",
+      "0x675b3dF06a5F3A7d2f04e7852Dbf8f8d40959Ca9",
+      "FRAX_Core",
       {
         base: "0",
-        multiplier: "0.15",
+        multiplier: "0.175",
         jump: "2.5",
         kink: "0.8",
       },
@@ -112,7 +114,19 @@ forking(8412432, async () => {
     );
 
     checkInterestRate(
-      "0x7a381Fa8B502ac3870D05ECfb321CFEf1818A218",
+      "0x675b3dF06a5F3A7d2f04e7852Dbf8f8d40959Ca9",
+      "DAI_CORE_TUSD_Core",
+      {
+        base: "0",
+        multiplier: "0.175",
+        jump: "2.5",
+        kink: "0.8",
+      },
+      BigNumber.from(ETH_BLOCKS_PER_YEAR),
+    );
+
+    checkInterestRate(
+      "0x786ed1931Fa76a70B1259c42B0aC106B1E08485c",
       "USDC_Ethena",
       {
         base: "0",
@@ -124,7 +138,7 @@ forking(8412432, async () => {
     );
   });
 
-  testForkedNetworkVipCommands("VIP 506", await vip506(), {
+  testForkedNetworkVipCommands("VIP 507", await vip507(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(txResponse, [VTOKEN_ABI], ["NewMarketInterestRateModel"], [9]);
     },
@@ -139,7 +153,7 @@ forking(8412432, async () => {
     });
 
     checkInterestRate(
-      SEPOLIA_vUSDS_Core_IRM,
+      ETH_vUSDS_Core_IRM,
       "USDS_Core",
       {
         base: "0",
@@ -151,7 +165,7 @@ forking(8412432, async () => {
     );
 
     checkTwoKinksInterestRateIL(
-      SEPOLIA_vUSDC_Core_IRM,
+      ETH_vUSDC_Core_IRM,
       "USDC_Core",
       {
         base: "0",
@@ -160,25 +174,25 @@ forking(8412432, async () => {
         multiplier2: "0.7",
         base2: "0",
         kink2: "0.9",
-        jump: "0.8",
+        jump: "3.0",
       },
       BigNumber.from(ETH_BLOCKS_PER_YEAR),
     );
 
     checkInterestRate(
-      SEPOLIA_vcrvUSD_Core_IRM,
+      ETH_vcrvUSD_Core_IRM,
       "crvUSD_Core",
       {
         base: "0",
         multiplier: "0.1",
-        jump: "0.8",
+        jump: "2.5",
         kink: "0.8",
       },
       BigNumber.from(ETH_BLOCKS_PER_YEAR),
     );
 
     checkInterestRate(
-      SEPOLIA_vFRAX_Core_IRM,
+      ETH_vFRAX_Core_IRM,
       "FRAX_Core",
       {
         base: "0",
@@ -190,7 +204,7 @@ forking(8412432, async () => {
     );
 
     checkInterestRate(
-      SEPOLIA_vDAI_Core_IRM,
+      ETH_vDAI_Core_IRM,
       "DAI_Core",
       {
         base: "0.",
@@ -202,7 +216,7 @@ forking(8412432, async () => {
     );
 
     checkTwoKinksInterestRateIL(
-      SEPOLIA_vUSDT_Core_IRM,
+      ETH_vUSDT_Core_IRM,
       "USDT_Core",
       {
         base: "0",
@@ -211,13 +225,13 @@ forking(8412432, async () => {
         multiplier2: "0.7",
         base2: "0",
         kink2: "0.9",
-        jump: "0.8",
+        jump: "3.0",
       },
       BigNumber.from(ETH_BLOCKS_PER_YEAR),
     );
 
     checkInterestRate(
-      SEPOLIA_vTUSD_Core_IRM,
+      ETH_vTUSD_Core_IRM,
       "TUSD_Core",
       {
         base: "0",
@@ -229,19 +243,19 @@ forking(8412432, async () => {
     );
 
     checkInterestRate(
-      SEPOLIA_vcrvUSD_Curve_IRM,
+      ETH_vcrvUSD_Curve_IRM,
       "crvUSD_Curve",
       {
         base: "0",
         multiplier: "0.1",
-        jump: "0.8",
+        jump: "2.5",
         kink: "0.8",
       },
       BigNumber.from(ETH_BLOCKS_PER_YEAR),
     );
 
     checkInterestRate(
-      SEPOLIA_vUSDC_Ethena_IRM,
+      ETH_vUSDC_Ethena_IRM,
       "USDC_Ethena",
       {
         base: "0",
