+++ simulations/vip-507/sepolia.ts
@@ -6,54 +6,55 @@ import { forking, testForkedNetworkVipCommands } from "src/vip-framework";
 import { checkInterestRate, checkTwoKinksInterestRateIL } from "src/vip-framework/checks/interestRateModel";
 
 import vip506, {
-  ETH_vDAI_Core,
-  ETH_vDAI_Core_IRM,
-  ETH_vFRAX_Core,
-  ETH_vFRAX_Core_IRM,
-  ETH_vTUSD_Core,
-  ETH_vTUSD_Core_IRM,
-  ETH_vUSDC_Core,
-  ETH_vUSDC_Core_IRM,
-  ETH_vUSDC_Ethena,
-  ETH_vUSDC_Ethena_IRM,
-  ETH_vUSDS_Core,
-  ETH_vUSDS_Core_IRM,
-  ETH_vUSDT_Core,
-  ETH_vUSDT_Core_IRM,
-  ETH_vcrvUSD_Core,
-  ETH_vcrvUSD_Core_IRM,
-  ETH_vcrvUSD_Curve,
-  ETH_vcrvUSD_Curve_IRM,
-} from "../../vips/vip-506/bscmainnet";
+  SEPOLIA_vDAI_Core,
+  SEPOLIA_vDAI_Core_IRM,
+  SEPOLIA_vFRAX_Core,
+  SEPOLIA_vFRAX_Core_IRM,
+  SEPOLIA_vTUSD_Core,
+  SEPOLIA_vTUSD_Core_IRM,
+  SEPOLIA_vUSDC_Core,
+  SEPOLIA_vUSDC_Core_IRM,
+  SEPOLIA_vUSDC_Ethena,
+  SEPOLIA_vUSDC_Ethena_IRM,
+  SEPOLIA_vUSDS_Core,
+  SEPOLIA_vUSDS_Core_IRM,
+  SEPOLIA_vUSDT_Core,
+  SEPOLIA_vUSDT_Core_IRM,
+  SEPOLIA_vcrvUSD_Core,
+  SEPOLIA_vcrvUSD_Core_IRM,
+  SEPOLIA_vcrvUSD_Curve,
+  SEPOLIA_vcrvUSD_Curve_IRM,
+} from "../../vips/vip-507/bsctestnet";
 import VTOKEN_ABI from "./abi/VToken.json";
 
 export const ETH_BLOCKS_PER_YEAR = 2_628_000; // assuming a block is mined every 12 seconds
+export const OLD_ETH_BLOCK_PER_YER = 2_252_571;
 
 const oldIRMs = [
-  [ETH_vUSDS_Core, "0x322072b84434609ff64333A125516055B5B4405F"],
-  [ETH_vUSDC_Core, "0x4A786e4653Ff7DBA74D6dA0861350F233f2dA73b"],
-  [ETH_vcrvUSD_Core, "0x837996e7d74222965ACf1fdd478926e07336a291"],
-  [ETH_vFRAX_Core, "0x675b3dF06a5F3A7d2f04e7852Dbf8f8d40959Ca9"],
-  [ETH_vDAI_Core, "0x675b3dF06a5F3A7d2f04e7852Dbf8f8d40959Ca9"],
-  [ETH_vUSDT_Core, "0x4A786e4653Ff7DBA74D6dA0861350F233f2dA73b"],
-  [ETH_vTUSD_Core, "0x675b3dF06a5F3A7d2f04e7852Dbf8f8d40959Ca9"],
-  [ETH_vcrvUSD_Curve, "0x837996e7d74222965ACf1fdd478926e07336a291"],
-  [ETH_vUSDC_Ethena, "0x786ed1931Fa76a70B1259c42B0aC106B1E08485c"],
+  [SEPOLIA_vUSDS_Core, "0xBd20E6922A5b1e20B5e611700f1Cab7c177C35De"],
+  [SEPOLIA_vUSDC_Core, "0xA03205bC635A772E533E7BE36b5701E331a70ea3"],
+  [SEPOLIA_vcrvUSD_Core, "0xA03205bC635A772E533E7BE36b5701E331a70ea3"],
+  [SEPOLIA_vFRAX_Core, "0xfBa27be0766acb9a60d2cede3d4293193f3b749d"],
+  [SEPOLIA_vDAI_Core, "0xB778B803a72Af7ac9Ac8277bA87B405cE91741B9"],
+  [SEPOLIA_vUSDT_Core, "0xA03205bC635A772E533E7BE36b5701E331a70ea3"],
+  [SEPOLIA_vTUSD_Core, "0xfBa27be0766acb9a60d2cede3d4293193f3b749d"],
+  [SEPOLIA_vcrvUSD_Curve, "0xA03205bC635A772E533E7BE36b5701E331a70ea3"],
+  [SEPOLIA_vUSDC_Ethena, "0x7a381Fa8B502ac3870D05ECfb321CFEf1818A218"],
 ];
 
 const newIRMs = [
-  [ETH_vUSDS_Core, ETH_vUSDS_Core_IRM],
-  [ETH_vUSDC_Core, ETH_vUSDC_Core_IRM],
-  [ETH_vcrvUSD_Core, ETH_vcrvUSD_Core_IRM],
-  [ETH_vFRAX_Core, ETH_vFRAX_Core_IRM],
-  [ETH_vDAI_Core, ETH_vDAI_Core_IRM],
-  [ETH_vUSDT_Core, ETH_vUSDT_Core_IRM],
-  [ETH_vTUSD_Core, ETH_vTUSD_Core_IRM],
-  [ETH_vcrvUSD_Curve, ETH_vcrvUSD_Curve_IRM],
-  [ETH_vUSDC_Ethena, ETH_vUSDC_Ethena_IRM],
+  [SEPOLIA_vUSDS_Core, SEPOLIA_vUSDS_Core_IRM],
+  [SEPOLIA_vUSDC_Core, SEPOLIA_vUSDC_Core_IRM],
+  [SEPOLIA_vcrvUSD_Core, SEPOLIA_vcrvUSD_Core_IRM],
+  [SEPOLIA_vFRAX_Core, SEPOLIA_vFRAX_Core_IRM],
+  [SEPOLIA_vDAI_Core, SEPOLIA_vDAI_Core_IRM],
+  [SEPOLIA_vUSDT_Core, SEPOLIA_vUSDT_Core_IRM],
+  [SEPOLIA_vTUSD_Core, SEPOLIA_vTUSD_Core_IRM],
+  [SEPOLIA_vcrvUSD_Curve, SEPOLIA_vcrvUSD_Curve_IRM],
+  [SEPOLIA_vUSDC_Ethena, SEPOLIA_vUSDC_Ethena_IRM],
 ];
 
-forking(22595258, async () => {
+forking(8412432, async () => {
   describe("Pre-VIP behaviour", async () => {
     it("check IRM address", async () => {
       for (const [market, expectedIRM] of oldIRMs) {
@@ -63,8 +64,8 @@ forking(22595258, async () => {
     });
 
     checkInterestRate(
-      "0x322072b84434609ff64333A125516055B5B4405F",
-      "USDS_Core",
+      "0xBd20E6922A5b1e20B5e611700f1Cab7c177C35De",
+      "USDS_CORE",
       {
         base: "0",
         multiplier: "0.15625",
@@ -74,27 +75,24 @@ forking(22595258, async () => {
       BigNumber.from(ETH_BLOCKS_PER_YEAR),
     );
 
-    checkTwoKinksInterestRateIL(
-      "0x4A786e4653Ff7DBA74D6dA0861350F233f2dA73b",
-      "USDC_Core_USDT_Core",
+    checkInterestRate(
+      "0xA03205bC635A772E533E7BE36b5701E331a70ea3",
+      "USDC_CORE_USDT_CORE_vcrvUSD_CORE_vcrv_USD_Curve",
+      {
+        base: "0",
+        multiplier: "0.07",
+        jump: "0.8",
+        kink: "0.8",
+      },
+      BigNumber.from(OLD_ETH_BLOCK_PER_YER),
+    );
+
+    checkInterestRate(
+      "0xfBa27be0766acb9a60d2cede3d4293193f3b749d",
+      "FRAX_Core_TUSD_Core",
       {
         base: "0",
         multiplier: "0.15",
-        kink1: "0.8",
-        multiplier2: "0.9",
-        base2: "0",
-        kink2: "0.9",
-        jump: "3.0",
-      },
-      BigNumber.from(ETH_BLOCKS_PER_YEAR),
-    );
-
-    checkInterestRate(
-      "0x837996e7d74222965ACf1fdd478926e07336a291",
-      "crvUSD_CORE_crvUSD_Curve",
-      {
-        base: "0",
-        multiplier: "0.2",
         jump: "2.5",
         kink: "0.8",
       },
@@ -102,11 +100,11 @@ forking(22595258, async () => {
     );
 
     checkInterestRate(
-      "0x675b3dF06a5F3A7d2f04e7852Dbf8f8d40959Ca9",
-      "FRAX_Core",
+      "0xB778B803a72Af7ac9Ac8277bA87B405cE91741B9",
+      "DAI_CORE",
       {
         base: "0",
-        multiplier: "0.175",
+        multiplier: "0.15",
         jump: "2.5",
         kink: "0.8",
       },
@@ -114,19 +112,7 @@ forking(22595258, async () => {
     );
 
     checkInterestRate(
-      "0x675b3dF06a5F3A7d2f04e7852Dbf8f8d40959Ca9",
-      "DAI_CORE_TUSD_Core",
-      {
-        base: "0",
-        multiplier: "0.175",
-        jump: "2.5",
-        kink: "0.8",
-      },
-      BigNumber.from(ETH_BLOCKS_PER_YEAR),
-    );
-
-    checkInterestRate(
-      "0x786ed1931Fa76a70B1259c42B0aC106B1E08485c",
+      "0x7a381Fa8B502ac3870D05ECfb321CFEf1818A218",
       "USDC_Ethena",
       {
         base: "0",
@@ -153,7 +139,7 @@ forking(22595258, async () => {
     });
 
     checkInterestRate(
-      ETH_vUSDS_Core_IRM,
+      SEPOLIA_vUSDS_Core_IRM,
       "USDS_Core",
       {
         base: "0",
@@ -165,7 +151,7 @@ forking(22595258, async () => {
     );
 
     checkTwoKinksInterestRateIL(
-      ETH_vUSDC_Core_IRM,
+      SEPOLIA_vUSDC_Core_IRM,
       "USDC_Core",
       {
         base: "0",
@@ -174,25 +160,25 @@ forking(22595258, async () => {
         multiplier2: "0.7",
         base2: "0",
         kink2: "0.9",
-        jump: "3.0",
+        jump: "0.8",
       },
       BigNumber.from(ETH_BLOCKS_PER_YEAR),
     );
 
     checkInterestRate(
-      ETH_vcrvUSD_Core_IRM,
+      SEPOLIA_vcrvUSD_Core_IRM,
       "crvUSD_Core",
       {
         base: "0",
         multiplier: "0.1",
-        jump: "2.5",
+        jump: "0.8",
         kink: "0.8",
       },
       BigNumber.from(ETH_BLOCKS_PER_YEAR),
     );
 
     checkInterestRate(
-      ETH_vFRAX_Core_IRM,
+      SEPOLIA_vFRAX_Core_IRM,
       "FRAX_Core",
       {
         base: "0",
@@ -204,7 +190,7 @@ forking(22595258, async () => {
     );
 
     checkInterestRate(
-      ETH_vDAI_Core_IRM,
+      SEPOLIA_vDAI_Core_IRM,
       "DAI_Core",
       {
         base: "0.",
@@ -216,7 +202,7 @@ forking(22595258, async () => {
     );
 
     checkTwoKinksInterestRateIL(
-      ETH_vUSDT_Core_IRM,
+      SEPOLIA_vUSDT_Core_IRM,
       "USDT_Core",
       {
         base: "0",
@@ -225,13 +211,13 @@ forking(22595258, async () => {
         multiplier2: "0.7",
         base2: "0",
         kink2: "0.9",
-        jump: "3.0",
+        jump: "0.8",
       },
       BigNumber.from(ETH_BLOCKS_PER_YEAR),
     );
 
     checkInterestRate(
-      ETH_vTUSD_Core_IRM,
+      SEPOLIA_vTUSD_Core_IRM,
       "TUSD_Core",
       {
         base: "0",
@@ -243,19 +229,19 @@ forking(22595258, async () => {
     );
 
     checkInterestRate(
-      ETH_vcrvUSD_Curve_IRM,
+      SEPOLIA_vcrvUSD_Curve_IRM,
       "crvUSD_Curve",
       {
         base: "0",
         multiplier: "0.1",
-        jump: "2.5",
+        jump: "0.8",
         kink: "0.8",
       },
       BigNumber.from(ETH_BLOCKS_PER_YEAR),
     );
 
     checkInterestRate(
-      ETH_vUSDC_Ethena_IRM,
+      SEPOLIA_vUSDC_Ethena_IRM,
       "USDC_Ethena",
       {
         base: "0",
