+++ simulations/vip-507/unichainsepolia.ts
@@ -5,24 +5,24 @@ import { expectEvents } from "src/utils";
 import { forking, testForkedNetworkVipCommands } from "src/vip-framework";
 import { checkInterestRate, checkTwoKinksInterestRateIL } from "src/vip-framework/checks/interestRateModel";
 
-import vip506, { BASE_vUSDC_Core, BASE_vUSDC_Core_IRM } from "../../vips/vip-506/bsctestnet";
+import vip506, { UNI_vUSDC_Core, UNI_vUSDC_Core_IRM } from "../../vips/vip-507/bsctestnet";
 import VTOKEN_ABI from "./abi/VToken.json";
 
-export const SECONDS_PER_YEAR = 31_536_000; // seconds per year
+export const SECONDS_PER_YEAR = 31_536_000;
 
-forking(26290932, async () => {
+forking(21502211, async () => {
   describe("Pre-VIP behaviour", async () => {
     it("check IRM address", async () => {
-      const vUSDC_Core = new ethers.Contract(BASE_vUSDC_Core, VTOKEN_ABI, ethers.provider);
-      expect(await vUSDC_Core.interestRateModel()).to.equals("0x37a3e7951795d3aF33D606A9591B8EBd5F84DA08");
+      const vUSDC_Core = new ethers.Contract(UNI_vUSDC_Core, VTOKEN_ABI, ethers.provider);
+      expect(await vUSDC_Core.interestRateModel()).to.equals("0x2f8263A8833C21b6B72AC55951756B41D4607e3F");
     });
 
     checkInterestRate(
-      "0x37a3e7951795d3aF33D606A9591B8EBd5F84DA08",
+      "0x2f8263A8833C21b6B72AC55951756B41D4607e3F",
       "USDC_CORE",
       {
         base: "0",
-        multiplier: "0.075",
+        multiplier: "0.0875",
         jump: "2.5",
         kink: "0.8",
       },
@@ -38,12 +38,12 @@ forking(26290932, async () => {
 
   describe("Post-VIP behavior", () => {
     it("check IRM address", async () => {
-      const vUSDC_Core = new ethers.Contract(BASE_vUSDC_Core, VTOKEN_ABI, ethers.provider);
-      expect(await vUSDC_Core.interestRateModel()).to.equals(BASE_vUSDC_Core_IRM);
+      const vUSDC_Core = new ethers.Contract(UNI_vUSDC_Core, VTOKEN_ABI, ethers.provider);
+      expect(await vUSDC_Core.interestRateModel()).to.equals(UNI_vUSDC_Core_IRM);
     });
 
     checkTwoKinksInterestRateIL(
-      BASE_vUSDC_Core_IRM,
+      UNI_vUSDC_Core_IRM,
       "USDC_Core",
       {
         base: "0",
